requirements.txt
fastapi
uvicorn
xgboost
scikit-learn
numpy
pandas
joblib
shap
lime
matplotlib
transformers
torch
accelerate
#huggingface-hub
#openai
groq
requests
python-dotenv
sentence-transformers
faiss-cpu
#tiktoken
#protobuf
